Fix compiler warnings
[Tracking bug 22690]

[git-p4: depot-paths = "//depot/software/common/ConsensusCore/": change = 127462]

+++ src/C++/Coverage.cpp
@@ -119,7 +119,7 @@ namespace ConsensusCore {
                              (endRowInChunk-startRowInChunk), tEnd+startRowInChunk,
                              chunkStart, CHUNK_SIZE, coverage);
             int j = 0;
-            while (j < chunkEnd - chunkStart)
+            while (j < (int)(chunkEnd - chunkStart))
             {
                 if (coverage[j] >= minCoverage)
                 {
